requirements.txt
requests
beautifulsoup4
PyYAML
selenium
webdriver-manager 